Initial url partial helper
Takes a url object as defined in config.yml as an argument.

If the url is external, it prints the url string that's a value in the object.

If the referenced URL is a member of a collection, this partial loops through the collection to find the permalink of the referenced item.

If given additional arguments, it will also inject a class that indicates the URL matches the current page's as well as inject arbitrary classes.